GameClassByFile: associates one file per unit, easing single view game display
We have a number of games which have data spread across various files.  For most games we require you to load each individual file in order to see the palettes there.  With GameClassByFile we now have support for associating each unit with a specific file, thus making a single view easy to implement.

Used this new class to provide KOF02UM stage support, as well as adding single views for each of the KOF02UM color format palette sets.

Also finished some remaining UINT16->size_t promotions, and added palette pairing for some Jojo's title screen sprites.

+++ palmod/Game/Game_DANKUGA_A.cpp
@@ -76,7 +76,7 @@ CGame_DanKuGa_A_DIR::CGame_DanKuGa_A_DIR(UINT32 nConfirmedROMSize)
 
     //Create the redirect buffer
     rgUnitRedir = new size_t[nUnitAmt + 1];
-    memset(rgUnitRedir, 0, sizeof(UINT16) * nUnitAmt);
+    memset(rgUnitRedir, 0, sizeof(size_t) * nUnitAmt);
 
     //Create the file changed flag
     PrepChangeTrackingArray();
